+++ src/card/cards.rs
@@ -11,6 +11,7 @@ mod prevention;
 mod rampage;
 mod regeneration;
 mod spore;
+mod targeted_regeneration;
 pub use attack_restrictions::*;
 pub use isd_rtr::*;
 pub use landwalk::*;
@@ -20,6 +21,7 @@ pub use prevention::*;
 pub use rampage::*;
 pub use regeneration::*;
 pub use spore::*;
+pub use targeted_regeneration::*;
 
 pub const ANKH_OF_MISHRA: CardDefinitionId = CardDefinitionId(1);
 pub const ATOG: CardDefinitionId = CardDefinitionId(2);